Unified all crates into a big one. The two binary crates (cmsis-cffi, cmsis-cli) remain separate.
