+++ rust/cmsis-pack/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod cmsis_update;
+pub mod pack_index;
+pub mod pdsc;
+pub mod utils;
