+++ rust/cmsis-pack/src/pack_index/mod.rs
@@ -1,6 +1,6 @@
 use minidom::{Element, Error};
-use slog::Logger;
-use utils::parse::{assert_root_name, attr_map, child_text, get_child_no_ns, FromElem};
+use slog::{Logger, o};
+use crate::utils::prelude::*;
 
 #[derive(Debug, Clone)]
 pub struct PdscRef {
