+++ rust/cmsis-pack/src/utils/prelude.rs
@@ -1,0 +1,11 @@
+pub use super::parse::{
+    attr_parse,
+    attr_parse_hex,
+    assert_root_name,
+    attr_map,
+    child_text,
+    get_child_no_ns,
+    FromElem
+};
+pub use super::ResultLogExt;
+pub use crate::err_msg;
