--- rust/cmsis-update/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "cmsis-update"
-version = "0.1.0"
-authors = ["Jimmy Brisson <theotherjimmy@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-futures = "0.1.25"
-minidom = "0.6.0"
-slog = "^2"
-slog-term = "^2"
-slog-async = "^2"
-tokio-core = "0.1.17"
-failure = "0.1.1"
-rustc-demangle = "=0.1.13"
-
-utils = { path = "../utils" }
-pack-index = { path = "../pack-index" }
-pdsc = { path = "../pdsc" }
-
-[dependencies.reqwest]
-version = "0.9.21"
-default_features = false
-features = ["rustls-tls"]
