--- rust/pack-index/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "pack-index"
-version = "0.1.0"
-authors = ["Jimmy Brisson <theotherjimmy@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-minidom = "0.6.0"
-quick-xml = "0.9.0"
-slog = "^2"
-failure = "0.1.1"
-utils = { path = "../utils" }
