--- rust/pdsc/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "pdsc"
-version = "0.1.0"
-authors = ["Jimmy Brisson <theotherjimmy@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-minidom = "0.6.0"
-slog = "^2"
-serde = "^1.0.89"
-serde_derive = "^1.0.89"
-serde_json = "1.0"
-failure = "0.1.1"
-utils = { path = "../utils" }
-pack-index = { path = "../pack-index" }
