--- rust/utils/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "utils"
-version = "0.1.0"
-authors = ["Jimmy Brisson <theotherjimmy@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-minidom = "0.6.0"
-quick-xml = "0.9.0"
-slog = "^2"
